test(ts): unify `Process Message` benchmark
